requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
deepdiff==7.0.1
distro==1.9.0
emoji==2.12.1
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.112.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.7
iopath==0.1.10
Jinja2==3.1.3
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.34
langchain-experimental==0.0.64
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.101
layoutparser==0.3.4
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
openai==1.42.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.2.0
pillow_heif==0.18.0
portalocker==2.10.1
protobuf==5.27.3
psutil==6.0.0
pycparser==2.22
pydantic
pydantic_core==2.20.1
pyparsing==3.1.2
pypdf==4.3.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rank-bm25==0.2.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.4
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
starlette==0.38.2
sympy==1.12
tabulate==0.9.0
tavily-python==0.4.0
tenacity==8.5.0
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.44.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.7
unstructured-client==0.25.5
unstructured-inference==0.7.36
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
wrapt==1.16.0
yarl==1.9.4